Change default locale help2man
`en` is not a valid locale, at least, not on my system (`locale -a`).
Because of unicode characters we prefer to use a unicode character set, like provided by the en_US.UTF8 locale, but without actually affecting other locale settings besides the character set.
This changes the default locale to `C.UTF-8` and falls back on `C` in case the former is unavailable. Furthermore, the environment variable HELP2MAN_LOCALE is provided to set another locale.